requirements.txt
# Core PyTorch Stack (Colab già ha torch preinstallato)
# torch>=2.0.0  # Commentato - usa versione Colab
# torchvision>=0.15.0 # Commented out to rely on Colab's pre-installed version
# torchaudio>=2.0.0

# Vision Backbones & Models
timm>=0.9.0  # PyTorch Image Models (ViT, ResNet, etc.)
torchmetrics>=1.0.0  # Required by DINOv3

# DINOv2 (loaded via torch.hub, no pip install needed)
# DINOv3 (not yet released publicly)
# SAM (optional, uncomment if needed):
# git+https://github.com/facebookresearch/dinov2.git  # Requires torch==2.0.0, uncomment if needed and install manually
# git+https://github.com/facebookresearch/dinov3.git  # May have torch version conflicts
# git+https://github.com/facebookresearch/segment-anything.git  # May have torch version conflicts


# Data Processing
# numpy>=1.23.0,<2.0.0  # IMPORTANTE: Commentato per evitare conflitti con Colab/TensorFlow/Numba
pillow>=9.0.0,<11.0.0
opencv-python>=4.8.0
scikit-image>=0.20.0

# Tensor Operations & Utilities
einops>=0.7.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration Management
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0

# Logging & Experiment Tracking
tensorboard>=2.13.0
wandb>=0.15.0

# Data Format Support
pandas>=1.5.0
scipy>=1.10.0

# Development Tools - VERSIONI COLAB
pytest>=7.3.0
black>=23.3.0
ruff>=0.0.270
# ipykernel - già installato in Colab
# jupyter-client - già installato in Colab

wandb>=0.15.0
tensorboard>=2.13.0